Update testfontsubstitution to new test file name conventions
